start
Update LICENSE.md

Create README.md

test

Added Trail class

Update README.md

Update README.md

COPYING

Update README.md

Added TrailList class

Edited Trail and TrailList class to add sorting functionality

Add database connect

Databaase connection is now working

Databaase connection is now working

database changes

user auth

Edited Trail and TrailList class to add sorting functionality

added method getTrailsByRegion, GetStates, GetRegions and GetCities

Added main activity page so that the user can select the states.  The list is empty right now because the data is on the database.

Added main activity page so that the user can select the states.  The list is empty right now because the data is on the database.

Added main activity page so that the user can select the states.  The list is empty right now because the data is on the database.

Add the cities classes and functionality for the UI.  The state and city classes have a temp list of items in it for now because the dbconnection's getCities and getStates were not returning data.

fixed cities and states

Added the regions class for the UI.  The arraylist has hardcoded data in it for now until the database is working.

Fixed up some of the Login screen so that the user doesn't have to input an email and MUST input a password.  The previous version allowed password to be optional and required an email.  This just made is to standard username and password with all of the correct error messages if field blank or password less than 5 chars.

Added a createSignIn class and UI so that the user can create an account.  The create an account is now hooked up to the sqlite database so that it stores the info.  I added correct error messages also.  I modified the login class so that the Sign in button appears and is functional.  The login class was also modified to check the database to see if the user name and password are in there and that they match.  I created a database class clased DataBaseHelper.  This has some functionality for the user login information.  I also changed the user class so that it matches the data needed for login and sign up.  I did minimal testing so far, so further testing of the classes is needed.

fixed cities and states

added trails

Fixed a bug in the trailListActivity.

added trails

Added Parcelable function to Trail so that the data can be moved through activities.  Changed the TrailListActivity class to start the TrailActivity class with the added passing of the Trail variable.

Added the trail information to the trail activity.

all styling

all styling

all styling

changed trails to add longitude and latitude and added some trails

added getters and setters for latitude and longitude in trails

google maps pluggin

Zooms in on position in the maps activity

Added more trails!!! Also changed the initialization of the dbconnection variables in trailListactivity and ThirdActivityRegions. For some reason it doesn't need to be initialized.

Added more trail.  Made TrailListActivity sort by ascending name so trails are easier to read.

Added a Sort By functionality in the trails list.  Changed TrailListACtivity and TrailList with the modifications.

sorting functionality

Update README.md

Presentation

Update build.gradle

Update activity_create_sign_up.xml

Update activity_create_sign_up.xml

copy write

copyright

copyright

Update activity_second_activity_cities.xml

copyright

copywrite

Update activity_trail_list.xml

copyright

copyright

copyright

Update strings.xml

copyright

copyright

copyright

copyright

copyright

copyright

Update ThirdActivityRegions.java

copyright

copyright

copyright

copyright

copyright

copyright

Update dbconnection.java

Changed license errors. Moved the license under the first line of code in .xml files

Update google_maps_api.xml